Extract demo tracker
